refactor: CardFormFields 폴더 내부 구조 변경
- 폴더 depth 줄이기
