+++ src/components/Card/CardFormFields/CardCvcInput.tsx
@@ -1,8 +1,8 @@
 import { ChangeEvent } from "react";
-import { CARD_FORM_TYPE } from "../../../../constants/constants";
-import Input from "../../../Common/Input/Input";
-import { useCard } from "../../../../hooks/useCard";
-import { CardValidationType } from "../../../../hooks/useCardValidation";
+import { CARD_FORM_TYPE } from "../../../constants/constants";
+import { useCard } from "../../../hooks/useCard";
+import { CardValidationType } from "../../../hooks/useCardValidation";
+import Input from "../../Common/Input/Input";
 
 const PLACEHOLDERS = "123";
 const MAX_LENGTHS = 3;
